pyproject.toml
[tool.poetry]
name = "rag-research"
version = "0.1.0"
description = "Testing and implementiation of a basic RAG for research purposes"
authors = ["Scott <scowty.sk@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
torch = "2.4.1"
transformers = "4.44.2"
huggingface-hub = "0.25.0"
black = "24.8.0"
flake8 = "7.1.1"
sentence-transformers = "3.1.1"
faiss-cpu = "*"
PyPDF2="3.0.1"
pdfplumber="0.11.4"
numpy="<2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
